Implemented missing iOS functions
Implemented isConnected for iOS, and since async functions don't behave the same
on iOS as Android. Implemented isSubbed and getTopics.

Will add async version is connect, disconnect, subscribe, unsubscribe

This will make it possible to have simple blocking functionality when trying to connect
to 10's or 100's of topics. Adding a hash table to look up the 'onConnect' event is
not simple to maintain across 4 client instances in the FE not to mention performant. It
is simpler to just make call and await the response or for a timeout. If either happens
then you can assume the client is hung